Refactor: constants 폴더 src 내부로 이동
